Assignment 3
For this project, a student registration system has been developed. The main functionality of the system is login to the student account, register a class, and view the student’s schedule. Moreover, the user can logout at the end of the process. Java Servlet, Java beans, JSP, and Derby database are the primary technology used.